Move decoder tests into source
